Initial commit
This code is garbage and will have to be rewritten.

When paired with the Asterisk setup, inbound calls to sip:some%40jid@server.tld
end up as working Jingle calls (tested only with Conversations) to the target
JID.

Declining the call, hanging up, and errors from calling bad JIDs all are handled.